refactor(consumption): extract LiveSampleSnapshot from TripRecordingController
Pull the "clock"-side snapshot out of TripRecordingController into a
focused collaborator: the per-PID latest-value scratch space, the
priority-tiered scheduler subscription wiring that fills it, and the
tier-1/2/3 fuel-rate derivation (plus the η_v resolution and fuel-trim
correction) that reads it.

The emit timer and `_emit` itself stay on the controller — that path
entangles the lifecycle flags, the recorder, and the fuel
accumulators. `_emit` now reads the snapshot once per tick. Scheduler
callbacks push high-priority parse outcomes and vehicle-speed samples
back through injected callbacks (the controller's silent-failure
observer and virtual-odometer buffer), so the collaborator carries no
drop-detection or distance state of its own.

Part of the "TripRecordingController split (clock)" acceptance
criterion of #1679.
